pdf_processor/requirements.txt
PyMuPDF
camelot-py[cv]
pandas
nltk
tiktoken
psycopg2-binary
pydantic
psutil
requests
pytesseract
Pillow
